+++ bcl/src/bc_radio.c
@@ -63,7 +63,7 @@ __attribute__((weak)) void bc_radio_on_thermometer(uint32_t *peer_device_address
 __attribute__((weak)) void bc_radio_on_humidity(uint32_t *peer_device_address, uint8_t *i2c, float *percentage) { (void) peer_device_address; (void) i2c; (void) percentage; }
 __attribute__((weak)) void bc_radio_on_lux_meter(uint32_t *peer_device_address, uint8_t *i2c, float *illuminance) { (void) peer_device_address; (void) i2c; (void) illuminance; }
 __attribute__((weak)) void bc_radio_on_barometer(uint32_t *peer_device_address, uint8_t *i2c, float *pressure, float *altitude) { (void) peer_device_address; (void) i2c; (void) pressure; (void) altitude; }
-__attribute__((weak)) void bc_radio_on_co2(uint32_t *peer_device_address, int16_t *concentration) { (void) peer_device_address; (void) concentration; }
+__attribute__((weak)) void bc_radio_on_co2(uint32_t *peer_device_address, float *concentration) { (void) peer_device_address; (void) concentration; }
 __attribute__((weak)) void bc_radio_on_buffer(uint32_t *peer_device_address, void *buffer, size_t *length) { (void) peer_device_address; (void) buffer; (void) length; }
 
 
@@ -216,7 +216,7 @@ bool bc_radio_pub_barometer(uint8_t i2c, float *pascal, float *meter)
     return true;
 }
 
-bool bc_radio_pub_co2(int16_t *concentration)
+bool bc_radio_pub_co2(float *concentration)
 {
     uint8_t buffer[1 + sizeof(*concentration)];
 
@@ -348,7 +348,7 @@ static void _bc_radio_task(void *param)
         }
         else if (queue_item_buffer[0] == BC_RADIO_HEADER_PUB_CO2)
         {
-            int16_t concentration;
+            float concentration;
 
             memcpy(&concentration, &queue_item_buffer[1], sizeof(concentration));
